+++ sdk/python/pulumi_alicloud/ecs/outputs.py
@@ -255,6 +255,7 @@ class InstanceDataDisk(dict):
                  encrypted: Optional[bool] = None,
                  kms_key_id: Optional[str] = None,
                  name: Optional[str] = None,
+                 performance_level: Optional[str] = None,
                  snapshot_id: Optional[str] = None):
         """
         :param int size: The size of the data disk.
@@ -276,6 +277,12 @@ class InstanceDataDisk(dict):
         :param bool encrypted: -(Optional, Bool, ForceNew) Encrypted the data in this disk. Default value: `false`.
         :param str kms_key_id: The KMS key ID corresponding to the Nth data disk.
         :param str name: The name of the data disk.
+        :param str performance_level: The performance level of the ESSD used as data disk:
+               - `PL0`: A single ESSD can deliver up to 10,000 random read/write IOPS.
+               - `PL1`: A single ESSD can deliver up to 50,000 random read/write IOPS.
+               - `PL2`: A single ESSD can deliver up to 100,000 random read/write IOPS.
+               - `PL3`: A single ESSD can deliver up to 1,000,000 random read/write IOPS.
+               Default to `PL1`.
         :param str snapshot_id: The snapshot ID used to initialize the data disk. If the size specified by snapshot is greater that the size of the disk, use the size specified by snapshot as the size of the data disk.
         """
         pulumi.set(__self__, "size", size)
@@ -293,6 +300,8 @@ class InstanceDataDisk(dict):
             pulumi.set(__self__, "kms_key_id", kms_key_id)
         if name is not None:
             pulumi.set(__self__, "name", name)
+        if performance_level is not None:
+            pulumi.set(__self__, "performance_level", performance_level)
         if snapshot_id is not None:
             pulumi.set(__self__, "snapshot_id", snapshot_id)
 
@@ -370,6 +379,19 @@ class InstanceDataDisk(dict):
         The name of the data disk.
         """
         return pulumi.get(self, "name")
+
+    @property
+    @pulumi.getter(name="performanceLevel")
+    def performance_level(self) -> Optional[str]:
+        """
+        The performance level of the ESSD used as data disk:
+        - `PL0`: A single ESSD can deliver up to 10,000 random read/write IOPS.
+        - `PL1`: A single ESSD can deliver up to 50,000 random read/write IOPS.
+        - `PL2`: A single ESSD can deliver up to 100,000 random read/write IOPS.
+        - `PL3`: A single ESSD can deliver up to 1,000,000 random read/write IOPS.
+        Default to `PL1`.
+        """
+        return pulumi.get(self, "performance_level")
 
     @property
     @pulumi.getter(name="snapshotId")