+++ sdk/python/pulumi_alicloud/ros/__init__.py
@@ -7,7 +7,9 @@ from .change_set import *
 from .get_change_sets import *
 from .get_stack_groups import *
 from .get_stacks import *
+from .get_templates import *
 from .stack import *
 from .stack_group import *
+from .template import *
 from ._inputs import *
 from . import outputs